advanced/requirements.txt
dnspython==2.7.0
Jinja2==3.1.6
python-nmap==0.7.1
requests==2.32.4
requests-file==2.1.0
shodan==1.31.0
theHarvester==0.0.1
wafw00f==2.3.1
whois==1.20240129.2
	
